requirements.txt
tensorflow>=2.13.0
pretty-midi>=0.2.9
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
